api/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.5.0
async-timeout==4.0.3
attrs==23.1.0
black==21.12b0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.6.2
distro==1.8.0
Django==4.0.1
django-cors-headers==3.11.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.7
frozenlist==1.4.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.334
langsmith==0.0.63
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.26.1
openai==0.28.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.3.0
python-dotenv==1.0.0
pytz==2021.3
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.2
tenacity==8.2.3
tomli==1.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
yarl==1.9.2
